1123/back/requirements.txt
asgiref==3.5.2
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
cryptography==38.0.3
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==3.2.13
django-allauth==0.50.0
django-cors-headers==3.13.0
django-extensions==3.2.0
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
drf-spectacular==0.24.2
executing==1.0.0
Faker==15.1.1
idna==3.4
inflection==0.5.1
ipython==8.4.0
jedi==0.18.1
jsonschema==4.17.0
matplotlib-inline==0.1.6
oauthlib==3.2.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.30
psycopg2==2.9.4
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyJWT==1.7.1
pyrsistent==0.19.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.5.0
toposort==1.7
traitlets==5.3.0
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.12
wcwidth==0.2.5
